+++ CHANGELOG.md
@@ -13,7 +13,8 @@ Changelog
 
 ### Fixes
 
-- [FIR, Anvil Interop] Fix rank-based binding replacements getting dropped for multi-contribution classes in root graphs.
+- [FIR, Anvil Interop] Fix rank-based binding replacements getting dropped for multi-contribution classes in root graphs when contributions are from external modules.
+- [IR, Anvil Interop] Fix rank-based binding replacements getting dropped for multi-contribution classes in graph extensions when contributions are from external modules.
 - [FIR] Named annotation arguments in different order from declared parameters getting silently skipped.
 
 0.9.2